# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PET - Portable Event Telephone API
  version: 1.0.0
  description: REST API for events, extensions, devices, DECT, callbacks and more.
paths:
...
  /api/v1/messaging/broadcast/:
    post:
      operationId: v1_messaging_broadcast_create
      description: '``{event: slug, text, group?: slug}`` -> orga text broadcast.'
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/messaging/messages/:
    get:
      operationId: v1_messaging_messages_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMessageList'
          description: ''
    post:
      operationId: v1_messaging_messages_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MessageRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MessageRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
...
components:
  schemas:
...
    DirectionEnum:
      enum:
      - out
      - in
      type: string
      description: |-
        * `out` - Outbound (PET → handset)
        * `in` - Inbound (handset → PET)
...
    Message:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
        direction:
          allOf:
          - $ref: '#/components/schemas/DirectionEnum'
          readOnly: true
        sender:
          type: string
          readOnly: true
        recipient:
          type: string
          readOnly: true
        text:
          type: string
          maxLength: 480
        state:
          allOf:
          - $ref: '#/components/schemas/MessageStateEnum'
          readOnly: true
        sent_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - direction
      - error
      - event
      - id
      - recipient
      - sender
      - sent_at
      - state
      - text
    MessageRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        recipient_number:
          type: string
          writeOnly: true
          minLength: 1
          maxLength: 16
        text:
          type: string
          minLength: 1
          maxLength: 480
      required:
      - event
      - recipient_number
      - text
    MessageStateEnum:
      enum:
      - queued
      - sent
      - failed
      - delivered
      type: string
      description: |-
        * `queued` - Queued
        * `sent` - Sent
        * `failed` - Failed
        * `delivered` - Delivered
...
    PaginatedMessageList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Message'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid